Grains: Remove uint12 type cast

+++ grains/grains.go
@@ -9,7 +9,7 @@ func Square(square int) (uint64, error) {
 	if square <= 0 || square > 64 {
 		return 0, errors.New("Input cannot be less than or equal to 0 or greater than 64")
 	}
-	return (1 << uint(square-1)), nil
+	return (1 << (square - 1)), nil
 }
 
 // Total returns the total number of grains on the chessboard.
